# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PR Reviewer Assignment Service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/pullRequest/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Создать PR и автоматически назначить до 2 ревьюверов из команды автора",
                "parameters": [
                    {
                        "description": "PR create payload",
                        "name": "pr",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreatePullRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/dto.PullRequest"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/merge": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Пометить PR как MERGED (идемпотентная операция)",
                "parameters": [
                    {
                        "description": "Merge request",
                        "name": "pr",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MergePullRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/dto.PullRequest"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/reassign": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Переназначить конкретного ревьювера на другого из его команды",
                "parameters": [
                    {
                        "description": "Reassign payload",
                        "name": "reassign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReassignPullRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PullRequestReassignResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Создать команду с участниками (создаёт/обновляет пользователей)",
                "parameters": [
                    {
                        "description": "Team object",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TeamAddInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/dto.Team"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/get": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Получить команду с участниками",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team name",
                        "name": "team_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Team"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/setIsActive": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Установить флаг активности пользователя",
                "parameters": [
                    {
                        "description": "User ID and active flag",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.setActiveInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/dto.User"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CreatePullRequest": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                }
            }
        },
        "domain.MergePullRequest": {
            "type": "object",
            "properties": {
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
        "domain.ReassignPullRequest": {
            "type": "object",
            "properties": {
                "old_reviewer_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
        "domain.TeamAddInput": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TeamAddMemberInput"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "domain.TeamAddMemberInput": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorObject": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorObject"
                }
            }
        },
        "dto.PullRequest": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "mergedAt": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.PullRequestReassignResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/dto.PullRequest"
                },
                "replaced_by": {
                    "type": "string"
                }
            }
        },
        "dto.Team": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TeamMember"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "dto.TeamMember": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.User": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "team_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.setActiveInput": {
            "type": "object",
            "required": [
                "is_active",
                "user_id"
            ],
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}